Update for SDL 6.0
* Hide classes
* Remove deprecated methods

+++ SmartDeviceLink/SDLUploadFileOperation.m
@@ -113,7 +113,7 @@ NS_ASSUME_NONNULL_BEGIN
         // The putfile's length parameter is based on the current offset
         SDLPutFile *putFile = [[SDLPutFile alloc] initWithFileName:file.name fileType:file.fileType persistentFile:file.isPersistent];
         putFile.offset = @(currentOffset);
-        putFile.length = @([self.class sdl_getPutFileLengthForOffset:currentOffset fileSize:file.fileSize mtuSize:mtuSize]);
+        putFile.length = @([self.class sdl_getPutFileLengthForOffset:currentOffset fileSize:(NSUInteger)file.fileSize mtuSize:mtuSize]);
 
         // Get a chunk of data from the input stream
         NSUInteger dataSize = [self.class sdl_getDataSizeForOffset:currentOffset fileSize:file.fileSize mtuSize:mtuSize];
@@ -177,7 +177,7 @@ NS_ASSUME_NONNULL_BEGIN
  @param mtuSize The maximum packet size allowed
  @return The the length of the data being sent in the putfile
  */
-+ (NSUInteger)sdl_getPutFileLengthForOffset:(NSUInteger)currentOffset fileSize:(unsigned long long)fileSize mtuSize:(NSUInteger)mtuSize {
++ (NSUInteger)sdl_getPutFileLengthForOffset:(NSUInteger)currentOffset fileSize:(NSUInteger)fileSize mtuSize:(NSUInteger)mtuSize {
     NSUInteger putFileLength = 0;
     if (currentOffset == 0) {
         // The first putfile sends the full file size
